gov_report/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gov-report"
version = "0.1.0"
description = "Government economic report crawler — fetch, convert, extract"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "tenacity>=8.2",
    "python-dateutil>=2.8",
    "beautifulsoup4>=4.12",
    "markdownify>=0.12",
    "feedparser>=6.0",
]

[project.scripts]
gov-report = "gov_report.cli:cli"

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src/gov_report"]
